Gesture event refactor
All gesture work moved into core from adaptor

+++ dali/integration-api/events/touch-event-integ.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014 Samsung Electronics Co., Ltd.
+ * Copyright (c) 2019 Samsung Electronics Co., Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -29,7 +29,7 @@ TouchEvent::TouchEvent()
 {
 }
 
-TouchEvent::TouchEvent( unsigned long time )
+TouchEvent::TouchEvent( uint32_t time )
 : MultiPointEvent( Touch, time )
 {
 }
